refactor(project): 规范后端代码并清理死代码
- 字符串散落改为枚举（生命周期动作/问题严重度状态/团队角色码）
- Controller业务逻辑下沉Service，返回类型强一致
- startReadiness万能Map改为StartReadinessResult
- 实体字段级注释与Service方法Javadoc补齐
- 枚举持久化统一为getCode()
- 展开通配符import、拆分超长行、格式化Controller与SQL脚本
- 清理businessType/businessId/countByWorkPackageId等死代码
- 16号阶段迁移脚本标注废弃

+++ ruoyi-project-management/src/main/java/com/ruoyi/projectmanagement/wbs/service/IProjectWbsService.java
@@ -1,12 +1,31 @@
 package com.ruoyi.projectmanagement.wbs.service;
+
 import com.ruoyi.projectmanagement.wbs.domain.ProjectWbsNode;
 import java.util.List;
+
+/**
+ * WBS范围树与工作包业务。
+ */
 public interface IProjectWbsService {
+
+    /** 查询WBS节点列表。 */
     List<ProjectWbsNode> list(ProjectWbsNode filter);
+
+    /** 查询WBS节点详细。 */
     ProjectWbsNode get(Long id);
-    int add(ProjectWbsNode node,String operator);
-    int edit(ProjectWbsNode node,String operator);
-    int remove(Long id,String operator);
+
+    /** 新增WBS节点，返回新节点ID。 */
+    Long add(ProjectWbsNode node, String operator);
+
+    /** 修改WBS节点。 */
+    int edit(ProjectWbsNode node, String operator);
+
+    /** 删除WBS节点。 */
+    int remove(Long id, String operator);
+
+    /** 刷新项目WBS汇总状态与进度。 */
     void refreshProject(Long projectId);
+
+    /** 判断项目是否所有工作包均已完成。 */
     boolean allWorkPackagesCompleted(Long projectId);
 }
